Folded SwarmCommands into existing CommandManager. Moved PathFinder to Ultils.
